Update from PR suggestions
