+++ Samples/Use multiple voucher codes on one order/Resources/Litium.Accelerator.Mvc/Site/Resources/ModuleECommerce/MultipleVoucherCodeCondition.sv-se.resx
